// frontend/src/shared/api/generated/models/eventStage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CodigoActivo.API
 * OpenAPI spec version: 1.0
 */

export type EventStage = typeof EventStage[keyof typeof EventStage];


export const EventStage = {
  Upcoming: 'Upcoming',
  EarlySignupOpen: 'EarlySignupOpen',
  SignupOpen: 'SignupOpen',
  SignupClosed: 'SignupClosed',
  Finished: 'Finished',
} as const;